change int to long

+++ src/Order.java
@@ -1,5 +1,5 @@
 public class Order {
-    private int count = 0;
+    private long count = 0;
 
     public Order() {
